+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Records can be skipped during import. Skipped records are recorded in result object.
 * Introduced BulkLoaderSource as a way of abstracting CSV / other source functionality away from the BulkLoader class.
 * Introduced ListBulkLoader for confining record CRUD actions to a given DataList (HasManyList).
+* Decoupled CSVParser from BulkLoader. Column mapping is now performed in BulkLoader on each record as it is loaded.
+* Replaced CSVParser with goodby/csv library.
 
 ## Bug Fixes